# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "Pet Diary",
    "contact": {}
  },
  "host": "127.0.0.1:3000",
  "basePath": "/api",
  "securityDefinitions": {},
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/user/update/profile": {
      "put": {
        "summary": "Update Profile",
        "tags": [
          "User"
        ],
        "operationId": "UpdateProfile",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/UpdateProfileRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
    "/pet/profil/updatepet": {
      "put": {
        "summary": "Update Profil Pet",
        "tags": [
          "Pet Data"
        ],
        "operationId": "UpdateProfilPet",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/UpdateProfilPetRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "PhotoPath": {
      "title": "PhotoPath",
      "example": {
        "photo1": "www.www"
      },
      "type": "object",
      "properties": {
        "photo1": {
          "type": "string"
        }
      },
      "required": [
        "photo1"
      ]
    },
...
    "UpdateProfileRequest": {
      "title": "UpdateProfileRequest",
      "example": {
        "uid_firebase": "1jffSKuCMuNjRuI6V1JlXRpEkay2",
        "full_name": "asengcoy",
        "phone": "082335506673",
        "photo_path": {
          "photo1": "wowowo.wowowow"
        },
        "email": "denandahp@gmail.com",
        "birthday": "1999-01-08"
      },
      "type": "object",
      "properties": {
        "uid_firebase": {
          "type": "string"
        },
        "full_name": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        },
        "photo_path": {
          "$ref": "#/definitions/PhotoPath"
        },
        "email": {
          "type": "string"
        },
        "birthday": {
          "type": "string"
        }
      },
      "required": [
        "uid_firebase",
        "full_name",
        "phone",
        "photo_path",
        "email",
        "birthday"
      ]
    },
...
    "UpdateProfilPetRequest": {
      "title": "UpdateProfilPetRequest",
      "example": {
        "uid_user": "1jffSKuCMuNjRuI6V1JlXRpEkay2",
        "pet_id": "4",
        "pet_name": "ipul ipul",
        "pet_born": "08-Jan-2015",
        "pet_gender": "MALE",
        "pet_bloodtype": "B",
        "pet_weight": "10",
        "pet_height": "20",
        "pet_type": "PERSIA",
        "pet_photo_path": "www.firestore"
      },
      "type": "object",
      "properties": {
        "uid_user": {
          "type": "string"
        },
        "pet_id": {
          "type": "string"
        },
        "pet_name": {
          "type": "string"
        },
        "pet_born": {
          "type": "string"
        },
        "pet_gender": {
          "type": "string"
        },
        "pet_bloodtype": {
          "type": "string"
        },
        "pet_weight": {
          "type": "string"
        },
        "pet_height": {
          "type": "string"
        },
        "pet_type": {
          "type": "string"
        },
        "pet_photo_path": {
          "type": "string"
        }
      },
      "required": [
        "uid_user",
        "pet_id",
        "pet_name",
        "pet_born",
        "pet_gender",
        "pet_bloodtype",
        "pet_weight",
        "pet_height",
        "pet_type",
        "pet_photo_path"
      ]
    },
...
  },
...
}